+++ NEWS.md
@@ -51,6 +51,8 @@
   when `na.action= "na.omit"`.
 - Overall, seven vignettes have been added to the package. All vignettes have a slightly
   different look now to reduce the size of the overall R package on CRAN.
+- The used optimizer is now available via `component(., "optimizer")` instead of previously
+  `attr(., "optimizer")`.
 
 # mmrm 0.2.2